admin role added and a new route to add the first admin user

+++ app/modules/users/routes.py
@@ -10,6 +10,13 @@ from app.db.engine import get_db
 # Typing Imports
 from typing import List
 
+# Exception imports
+from sqlalchemy.exc import IntegrityError
+
+# Authentication Imports
+from ..users.models import User
+from app.core.auth import manager
+
 # User Schemas
 from .services import UserService
 from .schemas import UserCreate
@@ -24,19 +31,25 @@ user_service = UserService()
 
 
 @route.get("/users/", response_model=List[UserResponse])
-def get_all_users(db: Session = Depends(get_db)):
+def get_all_users(db: Session = Depends(get_db), auth_user: User=Depends(manager)):
     """
     ## Retrieve a list of users.
+
+    ### Raises:  
+      >  HTTPException: Raises 401 is the user is not an admin.
 
     ### Returns:  
       >  List[UserResponse]: A List of users response models.
     """
+    if auth_user.admin == False:
+        raise HTTPException(status_code=401, detail="Access permitted only for admins")
+
     users = user_service.fetch_all(db)
     return users
 
 
 @route.get("/users/{id}", response_model=UserResponse)
-def get_one_user(id: int, db: Session = Depends(get_db)):
+def get_one_user(id: int, db: Session = Depends(get_db), auth_user: User=Depends(manager)):
     """
     ## Retrieve one user.
 
@@ -44,11 +57,15 @@ def get_one_user(id: int, db: Session = Depends(get_db)):
       >  id (int): The user ID.
 
     ### Raises:  
-      >  HTTPException: Raises 404 if user was not found.
+      >  HTTPException: Raises 404 if user was not found.  
+      >  HTTPException: Raises 401 is the user is not an admin.
 
     ### Returns:  
       >  UserResponse: The user response model.
     """
+    if auth_user.admin == False:
+        raise HTTPException(status_code=401, detail="Access permitted only for admins")
+
     user = user_service.fetch(db, id)
     if not user:
         raise HTTPException(status_code=404, detail="User was not found.")
@@ -56,35 +73,50 @@ def get_one_user(id: int, db: Session = Depends(get_db)):
 
 
 @route.post("/users/", response_model=UserResponse)
-def create_user(user: UserCreate, db: Session = Depends(get_db)):
+def create_user(user: UserCreate, db: Session = Depends(get_db), auth_user: User=Depends(manager)):
     """
     ## Creates an user.
 
     ### Args:  
-      >  user (UserCreate): The user update model.
+      >  user (UserCreate): The user model.
+
+    ### Raises:  
+      >  HTTPException: Raises 401 is the user is not an admin.  
+      >  HTTPException: Raises 422 if the email is already in use.  
 
     ### Returns:  
       >  UserResponse: The user response model.
     """
-    user = user_service.create(db, user)
-    return user
+    if auth_user.admin == False:
+        raise HTTPException(status_code=401, detail="Access permitted only for admins")
+
+    try:
+        user = user_service.create(db, user)
+        return user
+    except IntegrityError as err:
+        if "email" in repr(err):
+            raise HTTPException(status_code=422, detail="Já existe um usuário com este email cadastrado.")
 
 
 @route.patch("/users/{id}", response_model=UserResponse)
-def update_user(id: int, user: UserUpdate, db: Session = Depends(get_db)):
+def update_user(id: int, user: UserUpdate, db: Session = Depends(get_db), auth_user: User=Depends(manager)):
     """
     ## Edits an user by id.
 
     ### Args:  
       >  id (int): The user ID.  
-      >  user (UserUpdate): The user update model.
+      >  user (UserUpdate): The user model.
 
     ### Raises:  
-      >  HTTPException: Raises 404 if user was not found.
+      >  HTTPException: Raises 404 if user was not found.  
+      >  HTTPException: Raises 401 is the user is not an admin.
 
     ### Returns:  
       >  UserResponse: The user response model.
     """
+    if auth_user.admin == False:
+        raise HTTPException(status_code=401, detail="Access permitted only for admins")
+
     user = user_service.update(db, id, user)
     if not user:
         raise HTTPException(status_code=404, detail="User was not found.")
@@ -92,7 +124,7 @@ def update_user(id: int, user: UserUpdate, db: Session = Depends(get_db)):
 
 
 @route.delete("/users/{id}", response_model=UserResponse)
-def delete_user(id: int, db: Session = Depends(get_db)):
+def delete_user(id: int, db: Session = Depends(get_db), auth_user: User=Depends(manager)):
     """
     ## Deletes an user by id.
 
@@ -100,11 +132,15 @@ def delete_user(id: int, db: Session = Depends(get_db)):
       >  id (int): The user ID.
 
     ### Raises:  
-      >  HTTPException: Raises 404 if user was not found.
+      >  HTTPException: Raises 404 if user was not found.  
+      >  HTTPException: Raises 401 is the user is not an admin.
 
     ### Returns:  
       >  UserResponse: The user response model.
     """
+    if auth_user.admin == False:
+        raise HTTPException(status_code=401, detail="Access permitted only for admins")
+
     user = user_service.delete(db, id)
     if not user:
         raise HTTPException(status_code=404, detail="User was not found.")
